fix: fix TabsMobile width
