+++ packages/tabs/src/components/primary-tablist/Component.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useLayoutEffect, useRef } from 'react';
 import { KeyboardFocusable } from '@alfalab/core-components-keyboard-focusable';
 import cn from 'classnames';
 import { Styles, TabListProps } from '../../typings';
@@ -25,12 +25,12 @@ export const PrimaryTabList = ({
         onChange,
     });
 
-    useEffect(() => {
+    useLayoutEffect(() => {
         if (selectedTab && lineRef.current) {
             lineRef.current.style.width = `${selectedTab.offsetWidth}px`;
             lineRef.current.style.transform = `translateX(${selectedTab.offsetLeft}px)`;
         }
-    });
+    }, [selectedTab]);
 
     const renderContent = () => (
         <React.Fragment>
