+++ packages/tabs/src/components/tabs/Component.mobile/index.module.css
@@ -1,0 +1,3 @@
+.wrapper {
+    margin: 0 var(--gap-m-neg);
+}
