+++ packages/tabs/src/components/tabs/Component.mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './Component.mobile';
